+++ src/javascript/API.js
@@ -471,6 +471,13 @@ const graviton = {
       document.getElementById("g_dropmenu_list").style="";
       menus_showing=true;
     }
+  },
+  isInstalled: function(folder_name){
+    for(i=0;i<plugins_list.length;i++){
+      if(plugins_list[i].folder == folder_name){
+        return true;
+      }
+    }
   }
 }